Fixed an error with the recursive compression, and improved logging
